Disambiguate pow() call sincs dx() and dy() are ints.
svn path=/trunk/KDE/kdeedu/rocs/; revision=1178197

+++ src/Plugins/DSPlugins/Graph/OrientedEdgeItem.cpp
@@ -183,7 +183,7 @@ void OrientedEdgeItem::updatePos() {
     if( ! _edge ) remove();
     
     QLine q(_edge->from()->x(), _edge->from()->y(),    _edge->to()->x(),  _edge->to()->y());
-    qreal size = sqrt( pow(q.dx(), 2) + pow(q.dy(), 2));
+    qreal size = sqrt( pow(qreal(q.dx()), 2) + pow(qreal(q.dy()), 2));
     if (_edge->from() != _edge->to() && size < 20  ) {
         setPath(QPainterPath());
     } else {
