+++ src/Plugins/DSPlugins/LinkedList/PointerItem.cpp
@@ -146,7 +146,7 @@ void PointerItem::updatePos() {
         return;
     }
     QLine q(_edge->from()->x(), _edge->from()->y(),    _edge->to()->x(),  _edge->to()->y());
-    qreal size = sqrt( pow(q.dx(), 2) + pow(q.dy(), 2));
+    qreal size = sqrt( pow(qreal(q.dx()), 2) + pow(qreal(q.dy()), 2));
     if (_edge->from() != _edge->to() && size < 20  ) {
         setPath(QPainterPath());
     } else {
